cv-front | Removed image optimisation due to static needs
Next.js' image optimisation is on demand, server side.
The current web is static, so we will keep the control over images' size, and avoid this server cost.

If wanted, later on plugins can be added to optimise all images on build. Its setup is a little bit tricky, but can be done if needed. For now, it is not needed.

https://whereisthemouse.com/image-optimization-for-static-nextjs-sites