cv-front | Fixed colles form lang passing - WORKING
If lang is not destructured, it is passed as an Object and not as a string, and therefore it does not work.

Also updated Colles button to include language.
Also updated SuccessNotification and ErrorNotification at Create and Update colles to pass lang parameter before "reloading page" and losing language parameter, so that current language persists.

TODO 2.2: Add a language selector in a side bar.
TODO 3 (optional): Replace content.ts dictionary creation for csv sheet.