cv-front | Solved bug that disabled create/update colla form submit button when editing fields as if existing logo was invalid
Solved with a boolean check. Image content might have been somehow lost when redoing validations in fields other than the logo, as the handleLogoChange updated manually the image with its parameters.

Note: instead of simple boolean or let variables, in these javascript forms, initialised consts with a setter method have to be used in order for them to work.