cv-front | Saved as AVIF files to the remote server
In case of doubt, all images are stored as AVIF in the remote Ubuntu Server (to avoid previous possible JPEG compression due to saving them back as JPEG when a colla or figure was updated and a file had to be interpreted from Base64).

Still, inf the file format is different, it still has its original format when loaded back into the front.