cv-front | Added Colla Page
TODO 1: Solve URL parameter being lost when language is changed.
TODO 2: Stylise Colla Page

NOTE: Still, all usage of a colla relates to CollesContext (or for figura, FiguresContext), and as no id parameter has been found how to pass, always all colles or figures are retrieved. Could be improving this to optimise the web application in the future.