cv-front | User base card image added and stylisation started
TODO: Find proper style import strategy to avoid duplicating?
a) In general styles b) Importing from different modules c) Just duplicating

Note: to display images, the digital product with the active image is found based on ids, and then the image blob of the digital product is converted to an actual image.